analysis_bot/requirements.txt
fastapi
uvicorn
playwright
streamlit
google-generativeai
supabase
python-dotenv
pydantic
requests
pandas
